removed ;

+++ src/main/resources/init_db.sql
@@ -1,6 +1,6 @@
 CREATE DATABASE `taxi_db` /*!40100 DEFAULT CHARACTER SET utf8mb3 */ /*!80016 DEFAULT ENCRYPTION='N' */;
   CREATE TABLE `manufacturers` (
-    `id` int NOT NULL AUTO_INCREMENT,
+    `id` bigint NOT NULL AUTO_INCREMENT,
     `name` varchar(225) DEFAULT NULL,
     `country` varchar(225) DEFAULT NULL,
     `is_deleted` tinyint NOT NULL DEFAULT '0',
